implement album metadata with description
The gallery generator now rejects any files in an album that do not have
a .jpg, .gif, or .png extension (the only filetypes handled by
jekyll-image-tag) and tries to read a file called "album_info.y[a]ml" to
get metadata about an album. The metadata is merged into the album
page's "data" hash and thus made available to the page's template.

Currently the gallery generator does not do anything special with this
data, but it's useful for album descriptions, etc.

Fixes #6