sort: add filename sort to album metadata
It's sort of a no-op except for the reversing, but we can now specify
"filename" as the field to sort images on.

Also add in default metadata and a bit that checks _config.yml for
site metadata defaults before applying the data from the actual
album_info.yml file.

Fixes #7